Only uninstall gmp on Travis CI platform
[skip appveyor]

+++ recipe/build.sh
@@ -6,7 +6,10 @@ export CFLAGS="${CFLAGS} -O3 -I${PREFIX}/include"
 if [ "$(uname)" == "Darwin" ];
 then
     export LD_LIBRARY_PATH=$PREFIX/lib:$LD_LIBRARY_PATH
-    brew uninstall gmp
+    if [ $TRAVIS ];
+    then
+        brew uninstall gmp
+    fi
 fi
 
 ./configure --prefix=$PREFIX --with-gmp
